+++ src/pydocformatter/rules/definitions/PDF/PDF106_docstring_should_be_one_line.py
@@ -1,9 +1,11 @@
 from __future__ import annotations
 
 import pydocformatter.rules.collection as rule_collection
-from pydocformatter.rules.base import RuleBase, RuleCode, RuleMetadata
+from pydocformatter.rules.base import FixAvailability, RuleBase, RuleCode, RuleMetadata
 
 
 @rule_collection.register_rule
 class PDF106DocstringShouldBeOneLine(RuleBase):
-    meta = RuleMetadata(code=RuleCode("PDF106"), name="docstring-should-be-one-line", message="Docstring with one content line should be one line", fixable=True, stable_since="0.3.0")
+    meta = RuleMetadata(
+        code=RuleCode("PDF106"), name="docstring-should-be-one-line", message="Docstring with one content line should be one line", fix_availability=FixAvailability.ALWAYS, stable_since="0.3.0"
+    )